Remove overalignment from align < 8 allocations
This reorganizes the implementation of the System allocator to permit
adding various debuggig features. Currently, all that this implements is
a scheme that allocates a little extra space for low-alignment
allocations then returns a pointer into the actual allocation which is
offset so that it is not over-aligned.

This is a huge aid in discovering accidental reliance on over-alignment.
Allocators designed for C can be relied upon to produce over-aligned
pointers, so alignment-related bugs can be latent for a long time.

This implementation is also factored so accommodate other patches to the
default allocator which can help in detecting other sources of UB.